ppc/ppc400-telebot/requirements.txt
certifi==2016.8.31
DictObject==0.1.2
future==0.15.2
luckydonald-utils==0.48
peewee==2.8.3
Pillow==3.3.1
pkg-resources==0.0.0
pytg==0.4.10
python-telegram-bot==5.0.0
urllib3==1.17
